Rename `tag` to `id` (#45)
The `tag` is a concept used by the Notification API[1] that implies
uniqueness in a group, which will silently supersede any previous entity
having the same `tag` when it exists.

The `tag` in Background Fetch (and, unfortunately, Background Sync)
implies uniqueness in a group, but reuse will yield an exception instead.
This is a significant difference in behaviour.

Instead, switch to using `id` in Background Fetch. It still implies
uniqueness, but without countering the precedence set by notifications.